Refactor and fix so the map autoredraws on save
